requirements.txt
asgiref==3.5.0
async-timeout==4.0.3
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
cryptography==41.0.5
Django==4.0.3
django-cors-headers==4.3.0
djangorestframework==3.13.1
idna==3.4
pycparser==2.21
PyJWT==2.8.0
pytz==2021.3
redis==5.0.1
requests==2.31.0
setuptools==60.2.0
sqlparse==0.4.2
urllib3==2.1.0
wheel==0.37.1
gunicorn==21.2.0